Determine the winner of a round of rock, paper, scissors
I developed all the conditional if statements that
compares the user's choice and the computer's
choice and determine who won that round of "rock,
paper, scissors", if possible. Also, it will
display that the round was a draw when the user's
choice is the same as the computer's choice.

The next version of the JavaScript file will
include a function that adds a point to whoever
won the round of "rock, paper, scissors". Also,
the conditional statements will be consolidated
to a simplified function.