Allow the user to play a round of the game with the computer
I developed a function that takes the user's
selection and the computer's selection and plays
a round of "rock, paper, scissors". After the
function has been executed, the browser's console
will output a statement of how many point(s) the
user has and another statement of how many
point(s) the computer has.

The next version of the script will allow the user
to play 5 rounds of the game with the computer.
Also, the same statements of game points will be
displayed in the browser's console.